Fixed Image gradient overlay
Overlay would only be rendered properly in the Android Stuidoi Layout
inspector but neither on an emulator, nor on a real device.

Turns out, the problem was that the gradient was being passed as a simple view's
background and tehrefor needed a fixed width pre-runtime.
Moving the gradient as the background of a relative layout fixed the
problem.

Further reference:
http://stackoverflow.com/questions/20804564/overlay-over-an-imageview-on-android