Switched color vars from camelCase to snake_case
Refactroed three of the color variables defined by the theme. Tested in
the emulator and on a real device, both of which looked perfectly fine.

We should keep these changes in mind though, just in case something gets
messed up in the theme core later on, for this could be the reason for it.