Moved text colors to variables
For the alpha, there is some calculation that needs to be done.

A good guideline can be found
here: http://stackoverflow.com/a/11286013/4181679
I'll go through it on the example of the primary text, which has a color
of #000000 and a opacity of 87%.

First, we have to convert the 0.87 opacity into the alpha channel, which
ranges from 0 to 255. Therefore: 255 * 0.87 = 221,85 -> 222.
Then, we just convert this from base10 to base16 which is DE.
We then prepend this to our color and get #DE000000.