Main Activity Visual (#3)
* Refactored main activity and removed default actionbar

* Fixed package name for main activity

* Initialized basic toolbar in Main Activity

* Refactored Toolbar to be compatible with the menu drawer later on

* Added structure for drawables

* Added search icon

* Added XML for home menu

* Toolbar appearance finished

Toolbar itself is non-functional though.

* Set color of StatusBar to primeryDark

* Tidied up and befriended the compiler

* Basic card structure

* Styling for Series Cards done

Like the Menu, still non-functional

* Improved checkmark icon

* Switched Button to ImageButton

Also, added onClickListener in MainActivity

* Removed unnecessary imports

* Finished temporary Adapter for RecyclerView

* Added some Placeholder Material for Series Cards

* Fixed Image gradient overlay

Overlay would only be rendered properly in the Android Stuidoi Layout
inspector but neither on an emulator, nor on a real device.

Turns out, the problem was that the gradient was being passed as a simple view's
background and tehrefor needed a fixed width pre-runtime.
Moving the gradient as the background of a relative layout fixed the
problem.

Further reference:
http://stackoverflow.com/questions/20804564/overlay-over-an-imageview-on-android

* Fixed Layout for RecyclerView and Crads

* Added ripple effect and focus state

* Switched color vars from camelCase to snake_case

Refactroed three of the color variables defined by the theme. Tested in
the emulator and on a real device, both of which looked perfectly fine.

We should keep these changes in mind though, just in case something gets
messed up in the theme core later on, for this could be the reason for it.

* Moved text colors to variables

For the alpha, there is some calculation that needs to be done.

A good guideline can be found
here: http://stackoverflow.com/a/11286013/4181679
I'll go through it on the example of the primary text, which has a color
of #000000 and a opacity of 87%.

First, we have to convert the 0.87 opacity into the alpha channel, which
ranges from 0 to 255. Therefore: 255 * 0.87 = 221,85 -> 222.
Then, we just convert this from base10 to base16 which is DE.
We then prepend this to our color and get #DE000000.

* Moved onClickListener in Adapter

* Documentation and tidying up

* Merge master at 3744fccb659fc74f28f6df560395b39a39481e28

* Revert "Merge master at 3744fccb659fc74f28f6df560395b39a39481e28"

This reverts commit d1a2e664904e3d80c096ee9ad572897828be2986.

* Merge master at 3744fccb659fc74f28f6df560395b39a39481e28

* This merge should be correct to the line